Other functionalities Added

+++ src/main/java/com/nestaway/BeverageVendingMachine/Repository/BeverageRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BeverageRepo extends JpaRepository<BeverageModel, Integer> {
+public interface BeverageRepo extends JpaRepository<BeverageModel, Long> {
 }
