+++ src/main/java/com/nestaway/BeverageVendingMachine/Repository/InventoryRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InventoryRepo extends JpaRepository<InventoryModel, Integer> {
+public interface InventoryRepo extends JpaRepository<InventoryModel, Long> {
 }
